move funcs to utils
