+++ frontend/src/scenes/funnels/funnelUtils.ts
@@ -1,4 +1,5 @@
 import { FunnelStepReference } from 'scenes/insights/InsightTabs/FunnelTab/FunnelStepReferencePicker'
+import { getChartColors } from 'lib/colors'
 import { FunnelStep } from '~/types'
 
 export function calcPercentage(numerator: number, denominator: number): number {
@@ -18,4 +19,40 @@ export function getReferenceStep(steps: FunnelStep[], stepReference: FunnelStepR
         default:
             return steps[0]
     }
+}
+
+export function humanizeOrder(order: number): number {
+    return order + 1
+}
+
+export function getSeriesColor(index?: number): string | undefined {
+    if (typeof index === 'number' && index >= 0) {
+        return getChartColors('white')[index]
+    }
+    return
+}
+
+export function getBreakdownMaxIndex(breakdown?: FunnelStep[]): number | undefined {
+    // Returns the index of the last nonzero breakdown item
+    if (!breakdown) {
+        return
+    }
+    const nonZeroCounts = breakdown.map(({ count }, index) => ({ count, index })).filter(({ count }) => !!count)
+    if (!nonZeroCounts.length) {
+        return
+    }
+    return nonZeroCounts[nonZeroCounts.length - 1].index
+}
+
+export function getSeriesPositionName(
+    index?: number,
+    breakdownMaxIndex?: number
+): 'first' | 'last' | 'only' | undefined {
+    if (!breakdownMaxIndex) {
+        return 'only'
+    }
+    if (typeof index === 'number') {
+        return index === 0 ? 'first' : index === breakdownMaxIndex ? 'last' : undefined
+    }
+    return
 }
